refactor(tx-analyser): split into separate files and fix review findings
- Split TxAnalyser.tsx (1,272 lines) into 6 files under analyser/ directory
- Fix collectAddresses to collect both from and to addresses (DELEGATECALL enrichment)
- Fix CSS "monospace" to unquoted generic family name
